[ALPS07415365] Disp: Fix coverity issue
No conditions allow control to exit the loop.
ALPS07415365

always true
ALPS07414881

Dereference before null check
ALPS07415129
ALPS07415405

Macro compares unsigned to 0
ALPS07415022

UNREACHABLE
ALPS07415679

MTK-Commit-Id: eb7f2fd40d46805fefac64444c176fd78d9e7970

CR-Id: ALPS07415365
Feature: display_mobile

+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_drm_ddp_comp.c
@@ -1207,7 +1207,7 @@ void mtk_ddp_comp_clk_unprepare(struct mtk_ddp_comp *comp)
 
 	if (comp->mtk_crtc)
 		index = drm_crtc_index(&comp->mtk_crtc->base);
-	CRTC_MMP_MARK(index, ddp_clk, comp->id, 0);
+	CRTC_MMP_MARK((int) index, ddp_clk, comp->id, 0);
 }
 
 void mtk_ddp_comp_iommu_enable(struct mtk_ddp_comp *comp, struct cmdq_pkt *handle)
